refactor(ui): extract feed/skeleton.ts — FeedSkeletonItem placeholder component
